Change data storage to be std::byte everywhere.
This makes it clear everywhere that just raw data is being passed, not C-style strings, not a bunch of integers, just raw data.

+++ src/comms/USB/USBTransferContext.h
@@ -30,7 +30,7 @@ class USBTransferContext
 
 #ifdef __unix__
     libusb_transfer* transfer;
-    long bytesXfered;
+    int bytesXfered;
     std::atomic<bool> done;
     std::mutex transferLock;
     std::condition_variable cv;
